portfolai/requirements.txt
# Core Django
Django==5.2.7

# Django REST Framework (for API endpoints)
djangorestframework==3.15.2

# API integrations
openai>=1.50.2
finnhub-python==2.4.20
newsapi-python==0.2.7

# Environment management (for API keys)
python-dotenv==1.0.1

# CORS (if your frontend and backend communicate separately)
django-cors-headers==4.4.0

# For production deployments (optional)
gunicorn==23.0.0
whitenoise==6.6.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-timeout==2.3.1
coverage==7.3.2

# Celery for async task processing (optional)
# Uncomment when setting up Celery with Redis/RabbitMQ
# celery==5.3.4
# redis==5.0.1  # If using Redis as broker